HADOOP-5842.  Fix a few javac warnings under packages fs and util.  Contributed by Hairong Kuang
git-svn-id: https://svn.apache.org/repos/asf/hadoop/core/trunk@774980 13f79535-47bb-0310-9956-ffa450edef68

+++ src/core/org/apache/hadoop/fs/kfs/KFSInputStream.java
@@ -85,7 +85,7 @@ class KFSInputStream extends FSInputStream {
           if (statistics != null) {
             statistics.incrementBytesRead(1);
           }
-          return ((int) (b[0] & 0xff));
+          return b[0] & 0xff;
         }
         return -1;
     }
